+++ models/gamefront.go
@@ -1,15 +1,15 @@
 package models
 
 type GameFront struct {
-	GameId               *int                 `json:"GameId,omitempty"`
-	VendorId             *int                 `json:"VendorId,omitempty"`
+	GameId               *int64               `json:"GameId,omitempty"`
+	VendorId             *int64               `json:"VendorId,omitempty"`
 	VendorGameId         *string              `json:"VendorGameId,omitempty"`
 	VendorGameIdAlterate *string              `json:"VendorGameIdAlterate,omitempty"`
 	Name                 *string              `json:"Name,omitempty"`
 	ShortDescription     *string              `json:"ShortDescription,omitempty"`
 	Slug                 *string              `json:"Slug,omitempty"`
-	SuggestedHeight      *int                 `json:"SuggestedHeight,omitempty"`
-	SuggestedWidth       *int                 `json:"SuggestedWidth,omitempty"`
+	SuggestedHeight      *int64               `json:"SuggestedHeight,omitempty"`
+	SuggestedWidth       *int64               `json:"SuggestedWidth,omitempty"`
 	IsEnabled            *bool                `json:"IsEnabled,omitempty"`
 	IsNew                *bool                `json:"IsNew,omitempty"`
 	IsLive               *bool                `json:"IsLive,omitempty"`
@@ -23,8 +23,8 @@ type GameFront struct {
 	Vendor               *string              `json:"Vendor,omitempty"`
 	IsMinigame           *bool                `json:"IsMinigame,omitempty"`
 	IsParent             *bool                `json:"IsParent,omitempty"`
-	ParentGameId         *int                 `json:"ParentGameId,omitempty"`
-	ChildSortOrder       *int                 `json:"ChildSortOrder,omitempty"`
+	ParentGameId         *int64               `json:"ParentGameId,omitempty"`
+	ChildSortOrder       *int64               `json:"ChildSortOrder,omitempty"`
 	ChildGames           *[]GameFront         `json:"ChildGames,omitempty"`
 	ChildGameValue       *string              `json:"ChildGameValue,omitempty"`
 	InMaintenance        *bool                `json:"InMaintenance,omitempty"`
@@ -32,7 +32,7 @@ type GameFront struct {
 	Categories           *[]GameCategoryFront `json:"Categories,omitempty"`
 	Variants             *[]GameFrontVariant  `json:"Variants,omitempty"`
 	Jackpot              *GameJackpot         `json:"Jackpot,omitempty"`
-	OrderNumber          *int                 `json:"OrderNumber,omitempty"`
-	AdminCategoryId      *int                 `json:"AdminCategoryId,omitempty"`
+	OrderNumber          *int64               `json:"OrderNumber,omitempty"`
+	AdminCategoryId      *int64               `json:"AdminCategoryId,omitempty"`
 	IsLiveCasinoLobby    *bool                `json:"IsLiveCasinoLobby,omitempty"`
 }
